// day-5-react-routing-and-form/src/common/types.ts
export interface Task {
    id: number
    title: string
    completed: boolean
}

export interface TodosFormProps {
    addTask: (task: string) => void;
}

export interface TodosListProps {
    tasks: Task[]
    deleteTask: (e:React.MouseEvent, task: Task) => void;
    toggleTaskState: (task: Task) => void;
}

export interface HomeProps {
    taskList: Task[]
    setTaskList: React.Dispatch<React.SetStateAction<Task[]>>
}